MdeModulePkg: PciHostBridgeDxe: allow PciHostBridgeLib to prime the GCD map
Introduce the PCI_ROOT_BRIDGE.Flags member, which is a bitmap of 32 bits.
The least significant 5 bits are now assigned so that PciHostBridgeLib can
ask PciHostBridgeDxe *not* to add the various root bridge apertures to the
GCD IO and memory space map.

This is motivated by the case when some other module installs the same (or
larger) apertures in the GCD memory space map earlier, and
PciHostBridgeDxe's attempt to install overlapping entries would fail. For
example, the DXE core primes the GCD map from the IO and MMIO resource
descriptor HOBs that the HOB procuder (in practice: PEI) phase creates.

Cc: Ruiyu Ni <ruiyu.ni@intel.com
Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>