BaseTools: replace not defined macro to ' ' when generate response file
In last Nmake patch, when we generate the response file, we would replace
all the Macros in the make file. Once there have no definition macro
used, the tool direct report error. But for make tool, if there
have no definition macro, it will ignore this macro. so this patch change
the behavior to replace the not defined macro to ' ', but not report error.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Yonghong Zhu <yonghong.zhu@intel.com>