OvmfPkg: rename PcdAcpiPmBaseAddress to PcdPiix4AcpiPmBaseAddress
In the next patches, we'll differentiate the PMBA IO port address that we
program on PIIX4 vs. Q35.

Normally we'd just turn PcdAcpiPmBaseAddress into a dynamic PCD. However,
because we need this PCD in BaseRomAcpiTimerLib too (which cannot access
RAM and dynamic PCDs), it must remain a fixed PCD. We will introduce its
Q35 counterpart PCD later.

(We only need these PCDs to enable build-time configuration, and symbolic
references in the code. Introducing macros under
"OvmfPkg/Include/IndustryStandard/" is not a good choice, because the
values that the firmware programs are not standardized.)

As first step, rename the PCD so it says "Piix4" in the name, plus update
all client code to fetch it explicitly with FixedPcdGet16().

Cc: Jordan Justen <jordan.l.justen@intel.com>
Ref: https://bugzilla.redhat.com/show_bug.cgi?id=1333238
Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>