UefiCpuPkg/MpInitLib: allow platforms to provide a known CPU count upfront
On QEMU/KVM, the VCPU topology for a guest is specified dynamically. It
can be a low number or a high number.

Waiting for PcdCpuApInitTimeOutInMicroSeconds during the initial AP
collection is impractical, because in a VM, the time needed for an AP to
wake up can vary widely:

- if the APs report back quickly, then we could be wasting time,

- if the APs report back late (due to scheduling artifacts or VCPU
  over-subscription on the virtualization host), then the timeout can
  elapse before all APs increment CpuMpData->CpuCount in
  ApWakeupFunction().

Trying to set PcdCpuApInitTimeOutInMicroSeconds dynamically is also
impractical, as scheduling artifacts on the KVM host may delay AP threads
arbitrarily.

Instead, allow OVMF (and other platforms) to tell MpInitLib the number of
boot-time CPUs in advance.

Cc: Igor Mammedov <imammedo@redhat.com>
Cc: Jeff Fan <jeff.fan@intel.com>
Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>