ArmPkg: introduce EDKII Platform Has Device Tree Protocol
The presence of this protocol in the DXE protocol database implies that
the platform provides the operating system with a Device Tree-based
hardware description. This is not necessarily mutually exclusive with an
ACPI-based hardware description. A platform driver is supposed to produce
a single instance of the protocol (with NULL contents), if appropriate.

The decision to produce the protocol is platform specific; for example, it
could depend on an HII checkbox / underlying non-volatile UEFI variable.

The protocol is meant to be consumed by the same or another platform
driver that owns the Device Tree description of the hardware, and is
responsible for installing it as a system configuration table. Said
FDT-owner driver can wait for the protocol via DEPEX or protocol notify.

Because this protocol is not standard, it is prefixed with EDKII / Edkii,
as seen elsewhere in MdeModulePkg and SecurityPkg, for example. (ARM / Arm
doesn't look future-proof enough; future UEFI platforms could face the
same issue.)

Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Cc: Leif Lindholm <leif.lindholm@linaro.org>
Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>